refactor: break down into smaller components
